Update docs for required argument of `find`
`find()` has a required argument of type `KClass`, so `find<User>()` doesn't work since it has no arguments.

An alternative solution, of course, would be to change the `find()` function to match the docs if that is the desired behavior. I'm not sure if the `type` argument is used in a way that can't be done with just the generic `M`. If not, then the argument can just be removed in favor of the generic type. Otherwise I don't know an immediate solution offhand.